+++ commands/__init__.py
@@ -1,0 +1,9 @@
+from .create_aws_url_command import *
+from .ec2_info_command import *
+from .exit_command import *
+from .help_command import *
+from .remote_command_execute_command import *
+from .reset_command import *
+from .user_info_command import *
+from .user_privileges_command import *
+from .key import *
