+++ commands/exit_command.py
@@ -1,0 +1,7 @@
+import sys 
+from ishell.command import Command
+
+class ExitCommand(Command):
+    def run(self, line):
+        sys.exit()
+
