+++ commands/reset_command.py
@@ -1,0 +1,6 @@
+from .key import write_key 
+from ishell.command import Command
+
+class ResetCommand(Command):
+    def run(self, line):
+        write_key()
