+++ lib/__init__.py
@@ -1,0 +1,2 @@
+from .iam_client import IAMClient
+from .console import *
